requirements.txt
asgiref==3.11.0
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.1
django-dotenv==1.4.2
django-jazzmin==3.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
idna==3.11
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
